requirements.txt
requests==2.31.0
openpyxl==3.1.2
python-dateutil==2.8.2